+++ src/types/index.ts
@@ -70,6 +70,8 @@ export type {
   UploadActivityOptions,
   IntervalSearchOptions,
   ActivitiesAroundOptions,
+  Bucket,
+  TimeAtHRPlot,
   Interval,
   IntervalGroup,
   IntervalsDTO,
@@ -204,6 +206,9 @@ export type {
   ActivityHRCurve,
   ActivityHRCurvePayload,
   PaceDistancesDTO,
+  ActivityPowerCurvesOptions,
+  ActivityPaceCurvesOptions,
+  ActivityPaceCurves,
 } from './performance.js';
 
 // Fitness